pyproject.toml
[project]
name = "krakozhia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.86.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
